Add named script example for GitHub Actions
When executing the gh-pages command via a script
defined inside the packages.json file, it requires
the addition of an `--` parameter to properly pass
the user information into the `gh-pages` command.
A new example has been added reflecting that advice.

This commit also includes minor formatting fixes and
improvements the the existing GitHub Actions example.