fix: include dotfiles in the deployment by default
Whenever a user specifies a directory, they expect the dotfiles to be served as they exist on the file system. No web standards or conventions specifies otherwise. The previous default was surprising and as you see in the issue #315, has caused issues for many users.

This PR changes the default to true making this behaviour opt-in.